verifyページを切り離し
